getting prepared for Flyway 6.0
In Flyway 5.2.4, the configuration system was revamped and
the old one was deprecated and replaced by a fluent interface.
The old one will disappear in Flyway 6.0.

Error handler configuration was removed because we were initializing
it to an array of an empty string. If this had any meaning in the
old Flyway API, an alternative will have to be found using Callbacks.